+++ lib/output/async_writer.go
@@ -105,6 +105,7 @@ func (w *AsyncWriter) loop() {
 	)
 
 	defer func() {
+		w.writer.CloseAsync()
 		err := w.writer.WaitForClose(time.Second)
 		for ; err != nil; err = w.writer.WaitForClose(time.Second) {
 			w.log.Warnf("Waiting for output to close, blocked by: %v\n", err)