+++ lib/output/subprocess_test.go
@@ -1,0 +1,207 @@
+package output
+
+import (
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path"
+	"testing"
+	"time"
+
+	"github.com/Jeffail/benthos/v3/lib/log"
+	"github.com/Jeffail/benthos/v3/lib/message"
+	"github.com/Jeffail/benthos/v3/lib/metrics"
+	"github.com/Jeffail/benthos/v3/lib/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func testProgram(t *testing.T, program string) string {
+	t.Helper()
+
+	dir, err := ioutil.TempDir("", "benthos_subprocess_output_test")
+	require.NoError(t, err)
+
+	t.Cleanup(func() {
+		os.RemoveAll(dir)
+	})
+
+	pathStr := path.Join(dir, "main.go")
+	require.NoError(t, ioutil.WriteFile(pathStr, []byte(program), 0666))
+
+	return pathStr
+}
+
+func sendMsg(t *testing.T, msg string, tChan chan types.Transaction) {
+	t.Helper()
+
+	m := message.New(nil)
+	m.Append(message.NewPart([]byte(msg)))
+
+	resChan := make(chan types.Response)
+
+	select {
+	case tChan <- types.NewTransaction(m, resChan):
+	case <-time.After(time.Second):
+		t.Fatal("timed out")
+	}
+
+	select {
+	case res := <-resChan:
+		require.NotNil(t, res)
+		require.NoError(t, res.Error())
+	case <-time.After(time.Second):
+		t.Fatal("timed out")
+	}
+}
+
+func TestSubprocessBasic(t *testing.T) {
+	t.Parallel()
+
+	dir, err := ioutil.TempDir("", "benthos_subprocess_output_happy")
+	require.NoError(t, err)
+	t.Cleanup(func() {
+		os.RemoveAll(dir)
+	})
+
+	filePath := testProgram(t, fmt.Sprintf(`package main
+
+import (
+	"fmt"
+	"bufio"
+	"os"
+	"strings"
+	"bytes"
+	"io/ioutil"
+)
+
+func main() {
+	var buf bytes.Buffer
+
+	target := "%v/output.txt"
+
+	scanner := bufio.NewScanner(os.Stdin)
+	for scanner.Scan() {
+		fmt.Fprintln(&buf, strings.ToUpper(scanner.Text()))
+	}
+
+	if err := scanner.Err(); err != nil {
+		panic(err)
+	}
+
+	if err := ioutil.WriteFile(target, buf.Bytes(), 0644); err != nil {
+		panic(err)
+	}
+}
+`, dir))
+
+	conf := NewConfig()
+	conf.Type = TypeSubprocess
+	conf.Subprocess.Name = "go"
+	conf.Subprocess.Args = []string{"run", filePath}
+
+	o, err := New(conf, nil, log.Noop(), metrics.Noop())
+	require.NoError(t, err)
+
+	tranChan := make(chan types.Transaction)
+	o.Consume(tranChan)
+
+	sendMsg(t, "foo", tranChan)
+	sendMsg(t, "bar", tranChan)
+	sendMsg(t, "baz", tranChan)
+
+	o.CloseAsync()
+	o.CloseAsync() // No panic on double close
+
+	require.NoError(t, o.WaitForClose(time.Second))
+
+	assert.Eventually(t, func() bool {
+		resBytes, err := ioutil.ReadFile(path.Join(dir, "output.txt"))
+		if err != nil {
+			return false
+		}
+		return "FOO\nBAR\nBAZ\n" == string(resBytes)
+	}, time.Second, time.Millisecond*100)
+}
+
+func TestSubprocessEarlyExit(t *testing.T) {
+	t.Parallel()
+
+	dir, err := ioutil.TempDir("", "benthos_subprocess_output_early_exit")
+	require.NoError(t, err)
+	t.Cleanup(func() {
+		os.RemoveAll(dir)
+	})
+
+	filePath := testProgram(t, fmt.Sprintf(`package main
+
+import (
+	"fmt"
+	"bufio"
+	"os"
+	"strings"
+)
+
+func main() {
+	scanner := bufio.NewScanner(os.Stdin)
+	if scanner.Scan() {
+		f, err := os.OpenFile("%v/"+scanner.Text()+".txt", os.O_RDWR|os.O_CREATE, 0644)
+		if err != nil {
+			panic(err)
+		}
+		fmt.Fprintln(f, strings.ToUpper(scanner.Text()))
+		f.Sync()
+	} else if err := scanner.Err(); err != nil {
+		panic(err)
+	}
+}
+`, dir))
+
+	conf := NewConfig()
+	conf.Type = TypeSubprocess
+	conf.Subprocess.Name = "go"
+	conf.Subprocess.Args = []string{"run", filePath}
+
+	o, err := New(conf, nil, log.Noop(), metrics.Noop())
+	require.NoError(t, err)
+
+	tranChan := make(chan types.Transaction)
+	o.Consume(tranChan)
+
+	sendMsg(t, "foo", tranChan)
+
+	assert.Eventually(t, func() bool {
+		sendMsg(t, "bar", tranChan)
+
+		resBytes, err := ioutil.ReadFile(path.Join(dir, "bar.txt"))
+		if err != nil {
+			return false
+		}
+		return "BAR\n" == string(resBytes)
+	}, time.Second, time.Millisecond*100)
+
+	assert.Eventually(t, func() bool {
+		sendMsg(t, "baz", tranChan)
+
+		resBytes, err := ioutil.ReadFile(path.Join(dir, "baz.txt"))
+		if err != nil {
+			return false
+		}
+		return "BAZ\n" == string(resBytes)
+	}, time.Second, time.Millisecond*100)
+
+	o.CloseAsync()
+	require.NoError(t, o.WaitForClose(time.Second))
+
+	resBytes, err := ioutil.ReadFile(path.Join(dir, "foo.txt"))
+	require.NoError(t, err)
+	assert.Equal(t, "FOO\n", string(resBytes))
+
+	resBytes, err = ioutil.ReadFile(path.Join(dir, "bar.txt"))
+	require.NoError(t, err)
+	assert.Equal(t, "BAR\n", string(resBytes))
+
+	resBytes, err = ioutil.ReadFile(path.Join(dir, "baz.txt"))
+	require.NoError(t, err)
+	assert.Equal(t, "BAZ\n", string(resBytes))
+}